adding file structere and html
Added Basic File structure and full HTML for Personal
Homepage to be initially used as the Landing page prj
in the Foundations section of the Odin Project. Will cont
to update to make more personal touches, add JS and CSS,
and continue to build this as a living Personal Webpage and
Resume for current and future jobs.