Modify CSS to Flexbox Items
I went through and systematically corrected much
of the Flexbox issues, and scaled and sized everything
to fit within the bounds of the webpage. Will likely need
to adjust some flex shrink, scale items, but the bulk of
the webpage is functional. Will continue to edit and improve
as coding skills also improve.